Added Layback. - mainly for cab-owners.
Items on table are now displayed in a non-kartesian world to give the viewer the imagination to stand in front of the table instead of flying over it, when using FOV.

That was a tough one. Studied 3d programming over a week to know how to let this look good.

Tables only have to be slightly adjusted. Just set Inclination to 0, FOV to 45, layback to 45 and tweak scale and offsets so that the table fits.

I hope that this commit is complete. I've changed so many files.

git-svn-id: svn://svn.code.sf.net/p/vpinball/code/trunk@182 14438d79-0ffd-4333-8ae7-ea2bdd901387