fix very evil rdtsc abuse and increase timer precision a bit on the way
back in the days (i.e. ~2004) this was the way to go to get high precision timers, but nowadays the direct usage of rdtsc is full of sideeffects on multicore and speed throttling/turbo capable CPUs.

might fix a ton of strange issues ppl are seeing (long load times, micro stutter, initial fps issues), but then again, also maybe not.. ;)

git-svn-id: svn://svn.code.sf.net/p/vpinball/code/trunk@332 14438d79-0ffd-4333-8ae7-ea2bdd901387